be/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
requests==2.31.0
python-dotenv==1.0.0

# New for chatbot
openai==1.12.0
pinecone-client==3.0.0
beautifulsoup4==4.12.3
lxml==5.1.0
tiktoken==0.5.2
tenacity==8.2.3
